+++ packages/plugin-text/src/plugins/blockquote.tsx
@@ -1,0 +1,11 @@
+import { Editor } from 'slate'
+
+export const createBlockquote = (editor: Editor, name: string) => {
+  editor.command('replaceWithPlugin', {
+    plugin: 'blockquote',
+    state: {
+      plugin: name,
+      state: editor.value.toJSON()
+    }
+  })
+}