+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - **plugin-text**. Added list plugin
 - **plugin-text**. Added colors plugin
 - **plugin-text**. Added button to wrap in a blockquote plugin
+- **plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.
 - **plugin-rows**. Pass `remove` and `replace` functions to childs
 - **all**. Added plugin icons and descriptions